csm-streaming/requirements.txt
fastapi==0.117.1
Flask==3.1.2
huggingface_hub==0.35.0
librosa==0.11.0
matplotlib==3.10.6
moshi==0.2.2
numpy==2.3.3
psutil==7.1.0
psutil==7.0.0
pydantic==2.11.9
python-dotenv==1.1.1
Requests==2.32.5
safetensors==0.6.2
scikit_learn==1.7.2
scipy==1.16.2
sentence_transformers==5.1.0
sounddevice==0.5.0
SQLAlchemy==2.0.43
tokenizers==0.22.1
torch==2.8.0
torchaudio==2.8.0
torchtune==0.4.0
tqdm==4.67.1
transformers==4.56.1
typing_extensions==4.15.0
uvicorn==0.36.0
wandb==0.22.0
whisper_openai==1.0.0
